Moving server to src
